+++ auditlog/signals.py
@@ -20,7 +20,9 @@ Keyword arguments sent with this signal:
     audit log entry. Type: :class:`auditlog.models.LogEntry.Action`
 
 The receivers' return values are sent to any :func:`post_log`
-signal receivers.
+signal receivers, with one exception: if any receiver returns False,
+no logging will be made. This can be useful if logging should be
+conditionally enabled / disabled
 """
 
 post_log = django.dispatch.Signal()