+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### Improvements
 - feat: If any receiver returns False, no logging will be made. This can be useful if logging should be conditionally enabled / disabled ([#590](https://github.com/jazzband/django-auditlog/pull/590))
+- Django: Confirm Django 5.0 support ([#598](https://github.com/jazzband/django-auditlog/pull/598))
+- Django: Drop Django 4.1 support ([#598](https://github.com/jazzband/django-auditlog/pull/598))
 
 ## 3.0.0-beta.3 (2023-11-13)